Refactorized using make_shared and makeUniquePtr.
There's a bug in the communication channel for the levels, all the levels
inside level share the same eventQueue and GameEventDispatcher and the
events that happens inside a World is send to every other world.

The problem is that if a Ant Collide the action is repeated for every
World that the level has.